Addressed fire bug + wrong rotation on lasers
Hey Jasmeh I fixed the rotation error with a detailed explanation that I hope will help,

Also I took the liberty to fix the rapid fire error and also provided a description,

I added one more change with a comment. All the changes were made in the Enemy.cs script. If you have any questions feel free to send me an email or make an empty commit with a detailed description section, I'll try to pull the origin every few days or so to see if you have questions.

Enjoy the rest of your week :)