Changes to enemy pathing and enemy spawner
Changed the place where we call the yield return, it doesn't actually effect the logic we had earlier but It helps to reduce bugs of additional ships. I added some explanations I will go more in depth into on Friday,
I also rebuilt the script as Unity was getting missing files errors.

Made a change to the logic in enemy pathing so your gameplay loop continues. We can later combine this with our timer logic so our game feels better. and so we can properly keep spawning enemies after the timer has elapsed. (we will have to build out the coroutine later).

I will also go over recursion briefly as its what I used in this case. so this will be a sort of intro to data structures and algorithms.

Please make notes of anything you're not sure of so I can explain it on Friday.

Enjoy the rest of your week :)